backend/requirements.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
cryptography==45.0.3
dj-database-url==3.0.1
dj-rest-auth==7.0.1
django==5.2.1
django-allauth==65.3.0
django-appconf==1.1.0
django-auditlog==3.1.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-cryptography-5==2.0.3
django-filter==25.1
django-qstash==0.1.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
drf-spectacular==0.28.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
mailtrap==2.1.0
nh3==0.2.21
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
prompt-toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pycrypto==2.6.1
pyjwt==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-keyczar==0.716
pytz==2025.2
pyyaml==6.0.2
qstash==3.0.0
redis==5.2.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.25.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing-extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13

